+++ src/main/java/me/tomsdevsn/hetznercloud/objects/general/IPType.java
@@ -1,5 +1,6 @@
 package me.tomsdevsn.hetznercloud.objects.general;
 
-public enum TargetType {
-    label_selector,server
+public enum IPType {
+
+    ipv4,ipv6
 }