+++ src/main/java/me/tomsdevsn/hetznercloud/objects/request/FloatingIPRequest.java
@@ -3,6 +3,7 @@ package me.tomsdevsn.hetznercloud.objects.request;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
+import me.tomsdevsn.hetznercloud.objects.general.IPType;
 
 import java.util.Map;
 
@@ -12,7 +13,7 @@ import java.util.Map;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class FloatingIPRequest {
 
-    private String type;
+    private IPType type;
     @JsonProperty("home_location")
     private String homeLocation;
     private Long server;
@@ -20,18 +21,4 @@ public class FloatingIPRequest {
     private String name;
     @Singular
     private Map<String, String> labels;
-
-    @Getter
-    @AllArgsConstructor
-    public enum Type {
-        IPV4("ipv4"),
-        IPV6("ipv6");
-
-        private String type;
-
-        @Override
-        public String toString() {
-            return type;
-        }
-    }
 }